Repl

Description

Simple REPL library core, defines the REPL data type. Note: This module is not supposed to be imported! Instead, import String or Text

Documentation

data Repl a

Contains the REP operations of a REPL, and the loop breaking value.

Constructors

Repl

Fields

replRead :: IO a

Command used to read the input from the terminal.

replEval :: a -> a

Evaluation function.

replPrint :: a -> IO ()

Command used to print the result of evaluation.

replBreak :: a

Used to check if REPL should break the loop.

runRepl :: Eq a => Repl a -> IO ()

Starts the REPL

Note: Sets the buffering mode of <stdin> and <stdout> to NoBuffering.

If you alter the buffering mode during execution via the replRead or replPrint functions, the REPL will not work properly.